Short Range Enemy
Short range enemy is added, he is a grunt that walk left and right and is assigned a health value of 10. He should die from 1 player hit. Currently does not deal any damage to the player. Player Attacking is also added to this branch, and it works on left mouse button.